Spin out a ScreenshotConfiguration class to implement the configuration interface
